FILE: extracts createDto functions to separate files
