+++ libs/menu-matriarch/shared/data-access-dtos/src/index.ts
@@ -6,6 +6,14 @@ export * from './lib/menu-dto';
 export * from './lib/tag-dto';
 export * from './lib/user-dto';
 
+export * from './lib/creators/create-dish-dto';
+export * from './lib/creators/create-ingredient-dto';
+export * from './lib/creators/create-ingredient-type-dto';
+export * from './lib/creators/create-meal-dto';
+export * from './lib/creators/create-menu-dto';
+export * from './lib/creators/create-tag-dto';
+export * from './lib/creators/create-user-dto';
+
 export * from './lib/mappers/map-dish-dto-to-dish';
 export * from './lib/mappers/map-ingredient-type-dto-to-ingredient-type';
 export * from './lib/mappers/map-meal-dto-to-meal';
