+++ libs/menu-matriarch/shared/data-access-dtos/src/lib/creators/create-ingredient-dto.ts
@@ -1,0 +1,17 @@
+import { IngredientDto } from '../ingredient-dto';
+
+export function createIngredientDto({
+  id,
+  uid,
+  name,
+  typeId,
+  dishIds,
+}: Partial<IngredientDto>): IngredientDto {
+  return {
+    id: id ?? '',
+    uid: uid ?? '',
+    name: name ?? '',
+    typeId: typeId ?? '',
+    dishIds: dishIds ?? [],
+  };
+}
