+++ libs/menu-matriarch/shared/data-access-dtos/src/lib/creators/create-ingredient-type-dto.ts
@@ -1,0 +1,15 @@
+import { IngredientTypeDto } from '../ingredient-type-dto';
+
+export function createIngredientTypeDto({
+  id,
+  uid,
+  name,
+  ingredientIds,
+}: Partial<IngredientTypeDto>): IngredientTypeDto {
+  return {
+    id: id ?? '',
+    uid: uid ?? '',
+    name: name ?? '',
+    ingredientIds: ingredientIds ?? [],
+  };
+}
