+++ libs/menu-matriarch/shared/data-access-dtos/src/lib/creators/create-meal-dto.ts
@@ -1,0 +1,19 @@
+import { MealDto } from '../meal-dto';
+
+export function createMealDto({
+  id,
+  uid,
+  name,
+  description,
+  dishIds,
+  tagIds,
+}: Partial<MealDto>): MealDto {
+  return {
+    id: id ?? '',
+    uid: uid ?? '',
+    name: name ?? '',
+    description: description ?? '',
+    dishIds: dishIds ?? [],
+    tagIds: tagIds ?? [],
+  };
+}
