+++ libs/menu-matriarch/shared/data-access-dtos/src/lib/creators/create-menu-dto.spec.ts
@@ -1,4 +1,4 @@
-import { createMenuDto } from './menu-dto';
+import { createMenuDto } from './create-menu-dto';
 
 describe('createMenuDto', () => {
   it('creates default menu when no arguments are passed in', () => {
