+++ libs/menu-matriarch/shared/data-access-dtos/src/lib/creators/create-tag-dto.spec.ts
@@ -1,4 +1,4 @@
-import { createTagDto } from './tag-dto';
+import { createTagDto } from './create-tag-dto';
 
 describe('createTagDto', () => {
   it('creates default tag when no arguments are passed in', () => {
