+++ libs/menu-matriarch/shared/data-access-dtos/src/lib/ingredient-dto.ts
@@ -1,19 +1,3 @@
 import { Ingredient } from '@atocha/menu-matriarch/shared/util';
 
 export type IngredientDto = Ingredient;
-
-export function createIngredientDto({
-  id,
-  uid,
-  name,
-  typeId,
-  dishIds,
-}: Partial<IngredientDto>): IngredientDto {
-  return {
-    id: id ?? '',
-    uid: uid ?? '',
-    name: name ?? '',
-    typeId: typeId ?? '',
-    dishIds: dishIds ?? [],
-  };
-}
