+++ libs/menu-matriarch/shared/data-access-dtos/src/lib/ingredient-type-dto.ts
@@ -3,17 +3,3 @@ import { IngredientType } from '@atocha/menu-matriarch/shared/util';
 export interface IngredientTypeDto extends Omit<IngredientType, 'ingredients'> {
   ingredientIds: string[];
 }
-
-export function createIngredientTypeDto({
-  id,
-  uid,
-  name,
-  ingredientIds,
-}: Partial<IngredientTypeDto>): IngredientTypeDto {
-  return {
-    id: id ?? '',
-    uid: uid ?? '',
-    name: name ?? '',
-    ingredientIds: ingredientIds ?? [],
-  };
-}
