+++ libs/menu-matriarch/shared/data-access-dtos/src/lib/user-dto.ts
@@ -1,24 +1,3 @@
 import { User } from '@atocha/menu-matriarch/shared/util';
 
 export type UserDto = User;
-
-export function createUserDto({
-  uid,
-  name,
-  email,
-  preferences,
-}: Partial<UserDto>): UserDto {
-  return {
-    uid: uid ?? '',
-    name: name ?? '',
-    email: email ?? '',
-    preferences: {
-      darkMode: preferences?.darkMode ?? false,
-      dayNameDisplay: preferences?.dayNameDisplay ?? 'full',
-      defaultMenuStartDay: preferences?.defaultMenuStartDay ?? 'Monday',
-      emptyMealText: preferences?.emptyMealText ?? 'undecided',
-      mealOrientation: preferences?.mealOrientation ?? 'horizontal',
-      ingredientTypeOrder: preferences?.ingredientTypeOrder ?? [],
-    },
-  };
-}
